fixed. Сущность Room (DTO) - бекенд, создание, список.

+++ app/Modules/Auth/Application/Actions/Staff/RemoveStaffUseCase.php
@@ -5,6 +5,8 @@ namespace App\Modules\Auth\Application\Actions\Staff;
 use App\Modules\Auth\Application\Interfaces\StaffRepositoryInterface;
 use App\Modules\Shared\Domain\Entities\UserPermission;
 use App\Modules\Shared\Domain\Exceptions\AccessDeniedException;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 readonly class RemoveStaffUseCase
 {
@@ -13,10 +15,13 @@ readonly class RemoveStaffUseCase
     )
     {
     }
+
     public function execute(int $id, UserPermission $permissions): bool
     {
         if (!$permissions->can('auth.employee.delete')) throw new AccessDeniedException();
 
+        if (!$this->staffRepository->findById($id))
+            throw new NotFoundHttpException('Сотрудник не найден');
         //Проверка, можем ли удалить
 
         return $this->staffRepository->delete($id);
